Controller Route fixed
